requirements_personal.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.1
get==1.0.3
gunicorn==19.9.0
post==1.0.2
psycopg2==2.7.6.1
public==2.0.1
pytz==2018.7
query-string==1.0.2
request==1.0.2
static3==0.7.0
virtualenv==16.1.0
whitenoise==4.1.1
